+++ gentl-producer/src/misc/port.cpp
@@ -72,7 +72,7 @@ void Port::emitFeatureInvalidateEvent(const std::string& featureName) {
     if(featureInvalidateEvent != nullptr) {
         EventQueue::FeatureStringType buf; //MAX_LEN_FEATURE_NAME
         int len = featureName.length();
-        if (len >= sizeof(buf)) len = sizeof(buf)-1;
+        if (len >= (int)sizeof(buf)) len = sizeof(buf)-1;
         std::strncpy(buf.str, featureName.c_str(), len+1);
         buf.str[sizeof(buf)-1] = 0;
         DEBUG_PORT("Emitting (to queue) FEATURE_INVALIDATE for " << featureName << " -> DATA_ID=\"" << buf.str << "\"");
